// src/api/models/UpdateMatch.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateMatch
 */
export interface UpdateMatch {
    /**
     * 
     * @type {string}
     * @memberof UpdateMatch
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof UpdateMatch
     */
    matchDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof UpdateMatch
     */
    homeScore?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateMatch
     */
    awayTeamId?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateMatch
     */
    awayScore?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateMatch
     */
    homeTeamId?: number;
}

/**
 * Check if a given object implements the UpdateMatch interface.
 */
export function instanceOfUpdateMatch(value: object): boolean {
    return true;
}

export function UpdateMatchFromJSON(json: any): UpdateMatch {
    return UpdateMatchFromJSONTyped(json, false);
}

export function UpdateMatchFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateMatch {
    if (json == null) {
        return json;
    }
    return {
        
        'description': json['description'] == null ? undefined : json['description'],
        'matchDate': json['matchDate'] == null ? undefined : (new Date(json['matchDate'])),
        'homeScore': json['homeScore'] == null ? undefined : json['homeScore'],
        'awayTeamId': json['awayTeamId'] == null ? undefined : json['awayTeamId'],
        'awayScore': json['awayScore'] == null ? undefined : json['awayScore'],
        'homeTeamId': json['homeTeamId'] == null ? undefined : json['homeTeamId'],
    };
}

export function UpdateMatchToJSON(value?: UpdateMatch | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'description': value['description'],
        'matchDate': value['matchDate'] == null ? undefined : ((value['matchDate']).toISOString().substring(0,10)),
        'homeScore': value['homeScore'],
        'awayTeamId': value['awayTeamId'],
        'awayScore': value['awayScore'],
        'homeTeamId': value['homeTeamId'],
    };
}